avoid android error

+++ tests/google_tests/parser_test.cpp
@@ -250,9 +250,10 @@ TEST(ParserTest_, utf16escape) {
         int consumed;
 
         int utf16length = ParseUtil::parseUtf16toUtf8(str, strlen(str), 0, &consumed, &ch1, &ch2, &ch3, &ch4);
+
         EXPECT_EQ(utf16length, 2);
-        EXPECT_EQ((int)ch1, (int)0xC3);
-        EXPECT_EQ((int)ch2, (int)0x9F);
+        EXPECT_EQ(ch1, 0xC3);
+        EXPECT_EQ(ch2, 0x9F);
     }
 
     {
@@ -264,9 +265,9 @@ TEST(ParserTest_, utf16escape) {
         int consumed;
         int utf16length = ParseUtil::parseUtf16toUtf8(str, strlen(str), 0, &consumed, &ch1, &ch2, &ch3, &ch4);
         EXPECT_EQ(utf16length, 3);
-        EXPECT_EQ((int)ch1, (int)0xE8);
-        EXPECT_EQ((int)ch2, (int)0x99);
-        EXPECT_EQ((int)ch3, (int)0x8E);
+        EXPECT_EQ(ch1, 0xE8);
+        EXPECT_EQ(ch2, 0x99);
+        EXPECT_EQ(ch3, 0x8E);
     }
     
     {
@@ -282,15 +283,15 @@ TEST(ParserTest_, utf16escape) {
         int consumed;
         int utf16length = ParseUtil::parseUtf16toUtf8(str, strlen(str), 0, &consumed, &ch1, &ch2, &ch3, &ch4);
         EXPECT_EQ(utf16length, 4);
-        EXPECT_EQ((int)ch1, (int)0xF0);
-        EXPECT_EQ((int)ch2, (int)0xA0);
-        EXPECT_EQ((int)ch3, (int)0x8F);
-        EXPECT_EQ((int)ch4, (int)0xB9);
+        EXPECT_EQ(ch1, 0xF0);
+        EXPECT_EQ(ch2, 0xA0);
+        EXPECT_EQ(ch3, 0x8F);
+        EXPECT_EQ(ch4, 0xB9);
 
-        EXPECT_EQ((int)ch1, (int)rawStr[0]);
-        EXPECT_EQ((int)ch2, (int)rawStr[1]);
-        EXPECT_EQ((int)ch3, (int)rawStr[2]);
-        EXPECT_EQ((int)ch4, (int)rawStr[3]);
+        EXPECT_EQ(ch1, rawStr[0]);
+        EXPECT_EQ(ch2, rawStr[1]);
+        EXPECT_EQ(ch3, rawStr[2]);
+        EXPECT_EQ(ch4, rawStr[3]);
 
     }
